New species: steel elves
Wielding weapons in each hand, steel elves commune with the spirits
of ash and steel to guide their strikes. They take extra time to attune
to their weapons, meaning it takes as long for them to wield weapons as
it does for them to wear armour, and the animal magnetism granted to them
by the spirits often attracts newly seen enemies to them.

Slice and dice, or wield a staff in each hand. Go wild!

More notes on mechanics:
- Dual wielding: you can choose to wield a 1h weapon instead of a shield
  in your off-hand. (Unless you're wielding a 2h in your primary hand.)
  Whenever you strike with your primary weapon, your secondary will also
  attack. This doesn't trigger extra aux attacks. You can dual-wield 1h
  ranged weapons, but mixing and matching ranged and melee weapons doesn't
  work. Delay is the average of the two.
- Slow wielding: wielding, unwielding, or switching weapons (in both the
  primary and off-hand) takes 5 standard turns instead of 0.5.
- Attractivitis: When you first see a monster, there's a 1/3 chance of it
  being incredibly attracted to you. Such monsters are translocated to 2
  spaces away from you, as if you'd drunk a potion of attraction (but with
  no ongoing effects). This doesn't wake foes.

The idea is to play in a similar space to formicid's 2h+shield ability.
Dual wielding is expected to *usually* be stronger than wielding a 2h or
1h+shield - and if it isn't, we can buff it until it is - but gear
found will sometimes prompt one to use a shield or a 2h weapon, for a
shorter or longer period. That sort of 'soft' change for gear evaluation
(as opposed to a 'hard' lockout on item types ala spriggan/kobold/ogre/etc)
seems fun and exciting.

Attractivitis is intended to be a downside that pushes players more toward
engaging with the primary gimmick (melee weapons). Since dual wielding
changes strategic play, attractivitis is intended to make tactical play feel
more distinct. Putting a downside like this gives us more of a budget for
other kinds of power (apts, hp, etc) - and it's not always going to be a
downside, either!

More notes:
- Weapon names are purely cosmetic, but Esmeralda's feelings will be hurt if
  you say so.
- Apts and attributes are first drafts and especially subject to change.