New spell: Permafrost Eruption (L6 Ice/Earth)
It seems like it'd be fun to have an ice/earth spell. This one plays
with a targeting idea I've wanted to use for a while, an explosion
that automatically targets the largest clump of enemies onscreen.
(It'll never target anything adjacent to the player, which is a weakness.)

The spell does one physical hit to the whichever enemy is in the centre
of the blast, and another ice (half-rC) hit to that enemy and everything
adjacent to it. It's intended to encourage players to go into an unusual
school pairing by compensating for their school's weaknesses: Ice's
difficulty dealing with single targets, and Earth's issues with enemies
that have both EV and AC.

Damage numbers and Servitor support are TODOs.