Implement the North Wind (melee)
This wind (granted by the player's highest skill being UC or a melee
weapon, aside from short blades) causes 70% of the damage the player
sustains while rampaging to be converted into a new Indomitable status.
While Indomitable, the player's regeneration is massively increased (by a
rate dependant on their max HP) until all of this converted damage is
eventually recovered. In effect, this means that the player can shrug off
a majority of all damage they take while stampeding, so long as it isn't
happening too rapidly (or they are able to create a breather for
themselves).

This makes it so the player is often safer while performing
otherwise-reckless actions than they would be standing still, allowing
them to dive into enemy groups with relative impunity and pull off some
wild (and hopefully fun) maneuvers.

(I think the Indomitable status also has potential future applications as
a sort of 'delayed heal' that could be used in other contexts.